+++ CHANGELOG.md
@@ -9,6 +9,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * Added Notification for Session Changes
 
+### Added
+
+* Added support for adjusting the Session Replay sampling rate.
+
 ### Changed
 
 * The SDK now uses static binary dependencies to enhance integration in complex deployment scenarios.